# backend/openapi.json
{
  "components": {
    "schemas": {
      "Balance": {
        "properties": {
          "group_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Id"
          },
          "total_balance": {
            "title": "Total Balance",
            "type": "integer"
          },
          "total_owed_by_others": {
            "title": "Total Owed By Others",
            "type": "integer"
          },
          "total_owed_to_others": {
            "title": "Total Owed To Others",
            "type": "integer"
          },
          "user_balances": {
            "items": {
              "$ref": "#/components/schemas/UserBalance"
            },
            "title": "User Balances",
            "type": "array"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "total_balance",
          "total_owed_to_others",
          "total_owed_by_others",
          "user_balances"
        ],
        "title": "Balance",
        "type": "object"
      },
...
      "CreateGroup": {
        "properties": {
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "CreateGroup",
        "type": "object"
      },
...
      "Group": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "Group",
        "type": "object"
      },
      "GroupExpandedResponse": {
        "properties": {
          "balance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Balance"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "invites": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/GroupInviteResponse"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": [],
            "title": "Invites"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "transactions": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/TransactionRead"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": [],
            "title": "Transactions"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "users": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/UserResponse"
            },
            "title": "Users",
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "created_at",
          "updated_at"
        ],
        "title": "GroupExpandedResponse",
        "type": "object"
      },
...
      "GroupInviteResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "group_id": {
            "format": "uuid",
            "title": "Group Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "token": {
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "id",
          "group_id",
          "token",
          "created_at"
        ],
        "title": "GroupInviteResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TransactionParticipantRead": {
        "properties": {
          "amount_owed": {
            "title": "Amount Owed",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "debtor": {
            "$ref": "#/components/schemas/User"
          },
          "debtor_id": {
            "format": "uuid",
            "title": "Debtor Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "transaction_id": {
            "format": "uuid",
            "title": "Transaction Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "amount_owed",
          "transaction_id",
          "debtor_id",
          "id",
          "created_at",
          "updated_at",
          "debtor"
        ],
        "title": "TransactionParticipantRead",
        "type": "object"
      },
      "TransactionRead": {
        "properties": {
          "amount": {
            "description": "Total amount in smallest currency unit (e.g., cents)",
            "title": "Amount",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "group": {
            "$ref": "#/components/schemas/Group"
          },
          "group_id": {
            "format": "uuid",
            "title": "Group Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "participants": {
            "items": {
              "$ref": "#/components/schemas/TransactionParticipantRead"
            },
            "title": "Participants",
            "type": "array"
          },
          "payer": {
            "$ref": "#/components/schemas/User"
          },
          "payer_id": {
            "format": "uuid",
            "title": "Payer Id",
            "type": "string"
          },
          "purchased_on": {
            "format": "date-time",
            "title": "Purchased On",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "transaction_type": {
            "$ref": "#/components/schemas/TransactionType",
            "default": "EVEN"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "amount",
          "title",
          "group_id",
          "payer_id",
          "participants",
          "payer",
          "group"
        ],
        "title": "TransactionRead",
        "type": "object"
      },
      "TransactionType": {
        "enum": [
          "EVEN",
          "AMOUNT",
          "PERCENTAGE"
        ],
        "title": "TransactionType",
        "type": "string"
      },
...
      "UpdateGroup": {
        "properties": {
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "title": "UpdateGroup",
        "type": "object"
      },
      "User": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Email",
            "type": "string"
          },
          "email_verification_token": {
            "title": "Email Verification Token",
            "type": "integer"
          },
          "email_verified": {
            "default": false,
            "title": "Email Verified",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "password": {
            "minLength": 8,
            "title": "Password",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "username": {
            "maxLength": 50,
            "minLength": 1,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "email",
          "password"
        ],
        "title": "User",
        "type": "object"
      },
      "UserBalance": {
        "properties": {
          "balance": {
            "title": "Balance",
            "type": "integer"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          }
        },
        "required": [
          "user",
          "balance"
        ],
        "title": "UserBalance",
        "type": "object"
      },
...
      "UserResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "email_verified": {
            "title": "Email Verified",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "username",
          "email_verified",
          "created_at",
          "updated_at"
        ],
        "title": "UserResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/groups/": {
      "get": {
        "operationId": "read_groups_groups__get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Group"
                  },
                  "title": "Response Read Groups Groups  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Read Groups",
        "tags": [
          "groups"
        ]
      },
      "post": {
        "operationId": "create_group_groups__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroup"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Group"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Group",
        "tags": [
          "groups"
        ]
      }
    },
    "/groups/{group_id}": {
      "delete": {
        "operationId": "delete_group_groups__group_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "format": "uuid",
                  "type": "string"
                }
              ],
              "title": "Group Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete Group",
        "tags": [
          "groups"
        ]
      },
      "get": {
        "operationId": "read_group_groups__group_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "format": "uuid",
                  "type": "string"
                }
              ],
              "title": "Group Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupExpandedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Read Group",
        "tags": [
          "groups"
        ]
      },
      "put": {
        "operationId": "update_group_groups__group_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "format": "uuid",
                  "type": "string"
                }
              ],
              "title": "Group Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroup"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Group"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Group",
        "tags": [
          "groups"
        ]
      }
    },
    "/groups/{group_id}/transactions": {
      "get": {
        "operationId": "read_group_transactions_groups__group_id__transactions_get",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "format": "uuid",
                  "type": "string"
                }
              ],
              "title": "Group Id"
            }
          },
          {
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TransactionRead"
                  },
                  "title": "Response Read Group Transactions Groups  Group Id  Transactions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Read Group Transactions",
        "tags": [
          "groups",
          "transactions"
        ]
      }
    },
    "/groups/{group_id}/users/{user_id}": {
      "delete": {
        "operationId": "delete_user_from_group_groups__group_id__users__user_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "format": "uuid",
                  "type": "string"
                }
              ],
              "title": "Group Id"
            }
          },
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupExpandedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete User From Group",
        "tags": [
          "groups"
        ]
      }
    },
...
  }
}